avformat/mux: Remove pointless timestamp backups
write_packet currently saves the original timestamps of the packet it got
and restores them in case writing fails. This is unnecessary as none of
write_packet's callers make any use of these timestamps at all. So
remove this and add a general comment to the function that timestamps
may be modified; also remove a long outdated comment about side data.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>