avformat/avformat.h: Correct some comments
1. When set_parameters was removed from AVOutputFormat in 2fb75019, it
was forgotten to remove the comment pertaining to it. Said comment now
appeared to apply to interleave_packet; it is of course nonsense and has
been replaced by an accurate description.
2. The description of av_write_uncoded_frame suggested
av_interleaved_write_frame as a replacement if the input is not
already correctly interleaved; it also referred to said function for
details. Given that said function can't write AVFrames and that the
stuff specific to writing uncoded frames is explained in the description
of av_interleaved_write_uncoded_frame, both references have been fixed.
3. Removed an outdated comment about avformat_seek_file.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>