avformat/mux: Fix memleaks on errors II, improve documentation
1. When an error happened in ff_interleave_add_packet during
interleaving a packet for output, said packet would not be unreferenced
in ff_interleave_add_packet, but would be zeroed in
av_interleaved_write_frame, which results in a memleak.
2. When no error happened in ff_interleave_add_packet during
interleaving a packet for output, the input packet will already be blank
(as if it av_packet_unref had been applied to it), but it will
nevertheless be zeroed and initialized again. This is unnecessary.

Given the possibility of using other functions for interleavement than
ff_interleave_packet_per_dts, this commit sets and documents what
interleavement functions need to do: On success, the input packet (if
given) should be a blank packet on return. On failure, cleaning up will
be done by av_interleave_write_frame (where the already existing code
for error handling can be reused). ff_audio_rechunk_interleave has been
changed to abide by this. In order to keep these requirements
consistent, they are kept in one place that is referenced by the other
documentations.

Updating the documentation incidentally also removed another reference
to AVPacket's destructor.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>