avformat/matroskadec: Use already parsed cues if available
The WebM Dash specification [1] contains a requirement that a SeekHead
that references the Cues is mandatory if the Cues are after the
Clusters. There is no such requirement in case the Cues appear before
the Clusters.
Yet webm_dash_manifest_cues contained such an implicit requirement:
If it couldn't find an entry for Cues in the SeekHead, it concluded that
there were no Cues at all and returned an error.
This has been fixed: Now the list of all level 1 elements is searched for
Cues and said list contains already encountered Cues regardless of whether
they are referenced by a SeekHead.

[1]: http://wiki.webmproject.org/adaptive-streaming/webm-dash-specification

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>