avformat/matroskadec: Don't overlook parsing error
If one isn't in a cluster initially, matroska_parse_cluster would test
whether the next element is a cluster. Said test can fail and if it
failed, the error would be returned until 71c90881. But 71c90881 changed
this inadvertently: If the element found is an unknown-length element that
is intended to be skipped, but can't (because of its unknown length),
it is entered and if an error happens during parsing its children, it is
possible that one is on level 2 after the initial test. This will then
be treated as if one were inside a cluster and an attempt is made to
parse SimpleBlocks/BlockGroups in which case the original error will be
overwritten and overlooked. This commit fixes this.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>