avformat/matroskadec: Sanitize Cues
1. The Matroska CuePoints (the index entries) contain timestamps and
offsets as uint64_t. When adding them to a stream's index, they will be
automatically converted to int64_t. Up until now, there was no check for
whether a overflow happened during said conversion. This has been
changed. Given that values that overflow would be absurdly large, they
are treated as invalid data despite being in compliance with the
Matroska specifications.
2. A CuePoint has to contain a timestamp and at least a CueTrackPositions
which in turn needs to contain an offset. In order to check for the
existence of these elements, the timestamp as well as the offset now
have the default value UINT64_MAX, so that a CuePoint that doesn't contain
these elements will run afoul of the overflow checks mentioned above.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>