avformat/matroskadec: Use cues for subsegmentStartsWithSAP
1. According to the Dash specification [1] a subsegment is defined as
a unit within a Media Segment that is indexed by a Segment Index. There
may be more restrictions on what constitutes a subsegment and the WebM
Dash specification [2] makes use of this by imposing that a subsegment
must consist of one or more consecutive clusters.
It does not state that every cluster is a subsegment; yet this is how
the webm_dash_manifest_(de)muxer interpreted the specifications: The
demuxer seeked to the first position referenced by the cues and checked
whether it was the beginning of a cluster and whether the first packet
in said cluster was a keyframe. Then it seeked to the beginning of the
next cluster and checked it until it encountered something else than a
cluster or until it found a cluster that does not begin with a keyframe.
In the former case it added the metadata that all the clusters in the
stream begin with keyframes and the muxer would infer from this that it
can set subsegmentStartsWithSAP equal to 1 for this stream in the DASH
manifest.
This has been changed: Instead of checking every cluster, only the
clusters that are referenced in the cues are checked for whether they
begin with keyframes.
2. The earlier code also had the implicit assumption that the clusters
had a known length. The new method doesn't.
3. If the file ended with the last cluster, then the earlier code would
try to read past EOF which resulted in an error message. The new code
avoids this, too.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>