avcodec/h264_metadata: Add feature to repeat x264 SEI
x264 adds a user data unregistered SEI containing its version number and
encoding parameters to the first access unit. This SEI is mostly
informative, but it can also be used by a decoder (like FFmpeg's H.264
decoder) to decide whether to use workarounds for certain bugs in old
versions of x264. This in particular affects 4:4:4 content with the 8x8dct
option set. For such files it is imperative to know the x264 version and
there were scenarios where this information would be lost or not
gathered: The former can happen when the file is split, the latter can
happen when one doesn't decode from the very beginning.

Therefore this commit adds a feature to insert the x264 SEI in every
keyframe, but only to access units that don't already contain an x264
SEI. Changing x264 SEIs (which can be produced by concatenating x264
encodes) are supported.

Fixes ticket #8126.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>