fftools/ffmpeg: Reuse packet when possible during streamcopy
When do_streamcopy is called to output a packet, it must in general not
modify the packet, because the very same packet might also be output via
different OutputStreams. So it uses av_packet_ref() to create a new
reference to the packet's data. But this is unnecessary when the current
OutputStream is the last OutputStream that the packet is sent to (this
covers the common case in which a pkt is only sent to exactly one
OutputStream), because the packet will be unreferenced directly
afterwards anyway (in process_input()).

This commit implements a system in which the new references can be avoided
the last time a packet is sent to an OutputStream.

If the packet was refcounted initially, the old system had the
side-effect that the packet's data buffer was never writable. This meant
that certain bitstream filters (like truehd_core or noise) need to copy
the packet's data. This commit makes it possible to avoid these copies
in certain circumstances, e.g. if the packet was initially writable and
it is only sent to one OutputStream. (It is of course not possible to avoid
copying if the current OutputStream isn't the last OutputStream for that
packet; and a scenario where a stream gets both copied and routed through
a bitstream filter that wants the packet to be writable might not benefit
either because there might be a reference to the packet in the muxing
queue used for interleavement.)

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>