avformat/matroskaenc: Fix ReferenceBlock timestamp
In order to indicate that the frames in a BlockGroup are not keyframes,
one has to add a ReferenceBlock element containing the timestamp of a
reference block that has already been written. The timestamp ought to be
relative to the timestamp of the block it is attached to. Yet the
Matroska muxer used the relative timestamp of the preceding block of the
track, i.e. the timestamp of the preceding block relative to the
timestamp of the cluster containing said block (that need not be the
cluster containing the current block). This has been fixed.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>