avformat/matroskaenc: Allow empty filename for attachments
The EBML as well as the Matroska specifications allow empty strings, yet
the Matroska muxer enforced the requirement that a filename tag exists
for attachments and errored out if it didn't. This has been changed: In
the absence of a filename, an empty string is used.

That attached pictures (which ought to be written as AttachedFile by
this muxer, but aren't right now) from other containers than Matroska
likely don't have a filename metadata tag makes this commit a
prerequisite for writing attached-picture video streams as AttachedFile.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>