avformat/matroskaenc: Write level 1 elements in one go
Up until now, writing level 1 elements proceeded as follows: First, the
element id was written to the ordinary output AVIOContext and a dynamic
buffer was opened for the content of the level 1 element in
start_ebml_master_crc32(). Then this buffer was actually used and after it
was closed (in end_ebml_master_crc32()), the size field corresponding to
the buffer's size was written, after which the actual data was written.

This commit changes this: Nothing is written to the main AVIOContext any
more in start_ebml_master_crc32(). end_ebml_master_crc32() now writes
both the id, the size field as well as the data. This implies that
one can start a level 1 element in memory without outputting anything.
This is done in preparation for writing the attachments after the main
file data in case not all attachments are available when writing the
header (when the attachments are currently written).

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>