avformat/matroskaenc: Remove allocations for attachments
If there are attachments to write, the Matroska muxer currently
allocates two objects: An array that contains an entry for each
attachment containing just the stream index of the corresponding
stream and the uid used for this attachment; and a structure with
a pointer to said array and a counter for said array. These uids are
generated via code special to attachments: It uses an AVLFG in the
normal and a sha of the attachment data in the bitexact case. (Said sha
requires an allocation, too.)

But now that an uid is generated for each stream in mkv_init(), there is
no need any more to use special code for generating the uids of
attachments: One can simply use the uid already generated for the
attachment stream. And this makes the whole allocations of the structures
for attachments as well as the structures itself superfluous. They have
been removed.

In case AVFMT_FLAG_BITEXACT is set, the uids will be different from the
old ones which is the reason why the FATE-test lavf-mkv_attachment
needed to be updated. The old method had the drawback that muxing the
same attachment twice in the same file would create a file where two
attachments have the same uid. The new one doesn't.

Also notice that the dynamic buffer used to write the attachments leaks
if an error happens when writing the buffer. By removing the
allocations potential sources of errors have been removed.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>