avformat/matroskaenc: Add check for using explicit track number
When creating dash streams, the track number is externally prescribed
and not derived from the number of streams in the AVFormatContext. Up
until now, a track number of zero was allowed, although this is an
illegal track number.
Furthermore, it was not checked whether the number of tracks for a file
using an explicit track number was more than one, as such a file would
be invalid (it would be impossible to tell to which track a block belongs
if different tracks share the same track number).
Besides that, use the macro MAX_TRACKS macro for the maximum of
dash_track_number.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>