avformat/matroskaenc: Only write tracks if there is a track
The Matroska muxer does not write every stream as a Matroska track; some
streams are written as attachments. But should no stream be written as a
Matroska track, the Matroska muxer would nevertheless write a Tracks
element without a TrackEntry. This is against the spec. This commit
changes this and only writes a Tracks if there is a Matroska track.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>