avformat/matroskaenc: Automatically use right TrackNumber in cues
mkv_cuepoint (the structure used to store the index entries in our
Matroska muxer) currently contains fields for both the index of the
packet's stream in the AVFormatContext->streams array and for the
Matroska TrackNumber; correspondingly, mkv_add_cuepoint() has arguments
for both. But these two numbers can't be chosen independently, so get
rid of the TrackNumber.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>