avformat/matroskaenc: Improve writing video-projection
The Matroska video-projection master element has such a small maximum
length that it can always be written with a length field of length one.
So it is unnecessary to first write the element into a dynamic buffer to
get the accurate length in order not to waste bytes on the length field.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>