avcodec/cbs_mpeg2: Write slice data directly
Up until now, writing the data of a slice uses an intermediate buffer
into which the unit (both header as well as the rest) is assembled
before being copied into a freshly allocated buffer. But given that one
has a very good upper bound on the size needed before one starts copying
the slice data, one can allocate the buffer in advance, copy the already
written header into it and directly assemble the rest of the unit in
this buffer.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>