avcodec/cbs: Move allocation of unit data into write_unit
when writing a unit from content.

Writing a unit (i.e. assembling it from its decomposed form) currently
always uses an intermediate buffer; only after the unit has been
completely assembled in this buffer (so that, in particular, the needed
size is known), is the final buffer allocated and the data copied.

This works well for most types of units except one: The slice/frame
units that usually make up the bulk of the whole data. Only their
headers are reassembled from content. Their actual data are mostly just
copied into the write buffer. Given that the size of the actual data is
known, one has a very good estimate on the needed size after having
assembled the slice/frame header only. If one allocated the necessary
buffer now and directly copied the frame header and frame data into it,
one can avoid copying the data into the temporary buffer.

This commit lays the groundwork for doing so by moving the data buffer's
allocation into the (codec-dependent) write_unit function. It does not
avoid copying yet.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>