avformat/matroskaenc: Don't needlessly copy AVCodecParameters
At the end of encoding, the flac encoder sends a packet whose side-data
contains updated extradata (e.g. a correct md5 checksum). The Matroska
muxer uses this to update the CodecPrivate.

In doing so, the streams codecpar was copied. But given that writing
a flac CodecPrivate does not modify the used AVCodecParameters at all,
there is no need to do so and this commit changes this.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>