avformat/matroskaenc: Treat unknown duration correctly
For an AVPacket, a duration of zero means that the duration is unknown
(i.e. the end is marked by a stop-packet). Yet in Matroska, a duration
field of zero means exactly this: Display this block for 0 units of
time. Therefore it is wrong to map a duration of zero to a BlockGroup
with a duration element with value zero. In such a scenario it is better
to just not use a BlockGroup at all, but SimpleBlocks. (This is also what
mkvmerge does for subtitles of unknown duration.)

Furthermore, because a CueDuration of 0 makes no sense, apply the same
logic to cues and only write a CueDuration element if the duration does
not vanish.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>