avformat/matroskaenc: Remove special handling of WebVTT subs
Up until now, the WebM variant of WebVTT subtitles has been handled
specially: It had its own function to write it; this was so, because the
data had to be reformatted before writing. But given that other codecs
also need reformatting, this is no good reason to also duplicate the
generic stuff for writing Block(Group)s.

The reformatting used here uses a small buffer in order to avoid
allocations for the common case in which the subtitles are not too
large. This is similar to what avio_printf() does internally via an
AVBPrint.

This commit does not implement support for Matroska's way of muxing
WebVTT; it also does not add checks to ensure that WebM-style subtitles
don't get muxed in Matroska. But the function for reformatting gets a
webm prefix to indicate that this is for WebM.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>