avformat/matroskaenc: Sanitize timestamps
The current code did not handle timestamps < INT16_MIN very well: It
closed the current cluster (if one was open), opened a new cluster with
cluster timestamp 0 and wrote the timestamp as int16_t. This leads to
nonsense timestamps and also triggers an av_assert2 in avio_w8 (when
assert-level is big enough). This commit therefore clamps these
timestamps to INT16_MIN.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>