avformat/matroskaenc: Stop reallocating of cluster buffer
The Matroska muxer uses a dynamic buffer to buffer the content of
clusters before eventually writing them. Up until now, each time a
cluster was written, the dynamic buffer was closed and freed; now it is
only reset, saving allocations of the AVIOContext itself, its opaque
as well as most of the reallocations of the buffer.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>