avformat/matroskaenc: Improve cues in case of no-video
The Matroska muxer currently only adds CuePoints in three cases:
a) For video keyframes. b) For the first audio frame in a new cluster if
in dash-mode. c) For subtitles. This means that ordinary Matroska audio
files won't have any cues which impedes seeking.

This commit changes this. For every track in a file without video track
it is checked and tracked whether a cue entry has already been added
for said track for the current cluster. This is used to add a cue entry
for each first packet of each track in each cluster.

Implements #3149.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>