avformat/matroskaenc: Write attached_pic as attachment
This commit changes the way video streams with disposition
AV_DISPOSITION_ATTACHED_PIC that are not timed thumbnails and that
actually contain an attached_pic are handled: From now on, they will
be written as attachment, not as a Matroska Track.

The restriction to streams that actually contain an attached_pic is
important. E.g. simply setting the disposition via FFmpeg does not
populate the attached_pic field. Such tracks are treated as video tracks
just as now. It is intended to improve this in a future commit.

This fixes tickets #2270 and #4591.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>