avformat/matroskaenc: Don't waste bytes on BlockGroup length fields
The newly added helper functions for element sizes are now put to use to
calculate the exact length of the content of the BlockGroup as well as the
BlockAdditions and BlockMore master elements in advance so that the
least amount of bytes can be used for the length fields.

This fixes a regression introduced when the special code for writing
general subtitles was removed. Accordingly, the FATE-test binsub-mksenc
has now been reverted back to its old value again.

The new estimates are actually even sharper than the old ones, because
now the real size of the duration element and other uint/sint elements
are used, whereas only an upper bound was used before.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>