avformat/matroskadec: Don't keep the RealAudio pkt_cnt
pkt_cnt is always zero after every successfull call to
matroska_parse_rm_audio(); it can only be nonzero if there was not
enough memory for av_new_packet() or ff_packet_list_put(). If this
happens, the current code would try to output the remaining packets
whose data is in the current buffer when parsing the next block; but it
would also ignore the data contained in said block.

(Actually, the commit introducing this (eabb8ba4) did not check the
allocations at all and nevertheless kept this variable.)

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>