avformat/matroskadec: Set block index and total samples for WavPack
Each WavPack block in its native non-Matroska form contains a block
index (the sample number of the first sample of a given block) and a
field containing the number of samples of the entire file. Up until now
both of these fields were set to zero upon demuxing. This is wrong for
the number of total samples, as a value of UINT32_MAX is meant to
denote that the number of total samples is unknown. And it is wrong for
the block index, too, as this would mean that every block has a
timestamp of zero.

This commit changes this: The number of total samples is set to unknown;
and the block index is created by summing the number of samples output
so far; the Matroska timestamps are not used to synthesize the block
index, because the resolution of timestamps in Matroska is often not
sample-accurate.

The block index is already written in its new, 40-bit form (that utilizes
a previously unused byte in the header).

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>